Implemented additional assignment 2 logic, final program state

+++ Assignment2/ConsoleProcessor.h
@@ -2,8 +2,8 @@
 
 #ifndef CONSOLEPROCESSOR
 struct Coordinates {
-    int row;
-    int index;
+    size_t row;
+    ptrdiff_t index;
 };
 
 class Console {
